+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * [#299](https://github.com/TOMO-CAT/xmake/issues/299): 修复截断 package 的 commit version 导致无法拉取代码库的问题
 * [#301](https://github.com/TOMO-CAT/xmake/issues/301): 修复 force reinstall package 时安装失败的问题
 * [#312](https://github.com/TOMO-CAT/xmake/issues/312): 修复 cmake 错误识别 xmake config 携带的 cxxflags 参数
+* [#315](https://github.com/TOMO-CAT/xmake/issues/315): 修复 repo/package filelock 元数据写入锁文件导致 Linux 锁失效及并发反序列化失败的问题
 
 ## v3.0.8